Rename write.js to put.js
