+++ benchmarks/index.js
@@ -1,5 +1,5 @@
 'use strict'
 
-exports['write'] = require('./write')
+exports['put'] = require('./put')
 exports['batch-put'] = require('./batch-put')
 exports['self-distribution'] = require('./self-distribution')
